CBBF-86: Moved the exception handiling into the CustomSamplerClient based class.  Also added a handler for FhirClientConnectionException so it will come out as a 504 error in the jmeter log.  Renamed local runTest implementation to executeTest and made it pure virtual in the base class so tests can override it.  Also avoids confusion with the runTest used by the SamplerClient.
