+++ src/main/java/gov/hhs/cms/bluebutton/fhirstress/backend/CustomSamplerClient.java
@@ -2,10 +2,13 @@ package gov.hhs.cms.bluebutton.fhirstress.backend;
 
 import java.net.Inet4Address;
 import java.net.UnknownHostException;
+import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
 
 import org.apache.jmeter.config.Arguments;
 import org.apache.jmeter.protocol.java.sampler.AbstractJavaSamplerClient;
 import org.apache.jmeter.protocol.java.sampler.JavaSamplerContext;
+import org.apache.jmeter.samplers.SampleResult;
 
 import gov.hhs.cms.bluebutton.fhirstress.utils.FhirClient;
 //import gov.hhs.cms.bluebutton.fhirstress.utils.RifParser;
@@ -13,6 +16,8 @@ import gov.hhs.cms.bluebutton.fhirstress.utils.FhirClient;
 
 //import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.rest.client.IGenericClient;
+import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
+import ca.uhn.fhir.rest.client.exceptions.FhirClientConnectionException;
 
 public abstract class CustomSamplerClient extends AbstractJavaSamplerClient {
 	protected static final String PARAM_SERVER = "fhir_server";
@@ -22,6 +27,8 @@ public abstract class CustomSamplerClient extends AbstractJavaSamplerClient {
 	protected static final String RIFFILE = "riffile";
 	protected static final String DELIMITER = "delimiter";
 	protected static final String HOSTNAME_UNKNOWN = "unknown-host";
+
+  protected abstract void executeTest();
 
 	protected String hostName;
 	protected IGenericClient client;
@@ -71,6 +78,46 @@ public abstract class CustomSamplerClient extends AbstractJavaSamplerClient {
 		// Find the IDs that can be queried in each sample run.
 		//this.patientIds = findPatientIds();
 	}
+
+	/**
+	 * @see org.apache.jmeter.protocol.java.sampler.JavaSamplerClient#runTest(org.apache.jmeter.protocol.java.sampler.JavaSamplerContext)
+	 */
+	@Override
+	public SampleResult runTest(JavaSamplerContext context) {
+		SampleResult sample = new SampleResult();
+		sample.sampleStart();
+		try {
+			executeTest();
+
+			sample.sampleEnd();
+			sample.setSuccessful(true);
+			sample.setResponseMessage("Sample succeeded on host: " + hostName);
+			sample.setResponseCodeOK();
+		} catch (IllegalArgumentException|InvalidRequestException|FhirClientConnectionException e) {
+			// Mark this sample iteration as failed.
+			sample.sampleEnd();
+			sample.setSuccessful(false);
+			sample.setResponseMessage("Exception: " + e);
+
+			/*
+			 * Serialize the stack trace to a String and attach it to the sample
+			 * results.
+			 */
+			StringWriter stringWriter = new StringWriter();
+			e.printStackTrace(new java.io.PrintWriter(stringWriter));
+			sample.setResponseData(stringWriter.toString(), StandardCharsets.UTF_8.name());
+			sample.setDataType(org.apache.jmeter.samplers.SampleResult.TEXT);
+      if (e instanceof FhirClientConnectionException) {
+			  sample.setResponseCode("504");
+      }
+      else {
+			  sample.setResponseCode("500");
+      }
+		}
+
+		return sample;
+	}
+
 
 	/**
 	 * @return the hostname of the system this code is running on, or
